requirements.txt
appdirs==1.4.3
Appium-Python-Client==0.22
click==6.7
packaging==16.8
py==1.5.3
pyparsing==2.2.0
pytest==3.6.1
pytest-rerunfailures==2.1.0
python-dotenv==0.8.2
requests==2.10.0
selenium==3.4.2
six==1.10.0
